--- bin/Debug/net8.0/Calculator.deps.json
@@ -1,23 +1,0 @@
-{
-  "runtimeTarget": {
-    "name": ".NETCoreApp,Version=v8.0",
-    "signature": ""
-  },
-  "compilationOptions": {},
-  "targets": {
-    ".NETCoreApp,Version=v8.0": {
-      "Calculator/1.0.0": {
-        "runtime": {
-          "Calculator.dll": {}
-        }
-      }
-    }
-  },
-  "libraries": {
-    "Calculator/1.0.0": {
-      "type": "project",
-      "serviceable": false,
-      "sha512": ""
-    }
-  }
-}
